Feat: Let user specify a user agent prefix
This patch exposes an option in `SessionOptions` that a user can set,
which will be prefixed to the beginning of the user agent sent to the
broker during session negotiation.  This prefix does not need to be
set, but if it is, it must consist only of printable ASCII characters
and must be less than 128 characters long.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>