Fix: avoid implicit cast in compound assignment
This patch fixes a warning on newer JDKs, where we narrow a `long`
down to an `int` during a compound assignment later on in the
function.  This `long` is the result of the `readByte()` call on the
line changed by this patch, and we might as well store it as an `int`
early on.

    [WARNING] /Users/pniedzielski/Projects/bloomberg/blazingmq-sdk-java/bmq-sdk/src/main/java/com/bloomberg/bmq/impl/infr/proto/MessagePropertiesImpl.java:[234,24] implicit cast from long to int in compound assignment is possibly lossy

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>